[K6.0] Error in the template manager - Actions are not executed #8419
