+++ src/admin/src/Controller/TemplateController.php
@@ -24,6 +24,7 @@ use Joomla\CMS\Session\Session;
 use Joomla\Registry\Registry;
 use Joomla\Utilities\ArrayHelper;
 use Kunena\Forum\Libraries\Factory\KunenaFactory;
+use Kunena\Forum\Libraries\Path\KunenaPath;
 use Kunena\Forum\Libraries\Route\KunenaRoute;
 
 /**
@@ -57,7 +58,7 @@ class TemplateController extends FormController
 	{
 		parent::__construct($config, $factory, $app, $input);
 
-		$this->baseurl = 'administrator/index.php?option=com_kunena&view=templates';
+		$this->baseurl = 'administrator/index.php?option=com_kunena&view=template';
 	}
 
 	/**
@@ -384,4 +385,87 @@ class TemplateController extends FormController
 		$this->setRedirect(KunenaRoute::_($this->baseurl, false));
 	}
 
+	/**
+	 * Edit the template selected
+	 *
+	 * @param   null  $key     key
+	 * @param   null  $urlVar  url var
+	 *
+	 * @return  void
+	 *
+	 * @throws  Exception
+	 * @since   Kunena 2.0
+	 */
+	public function edit($key = null, $urlVar = null)
+	{
+		$cid      = $this->app->input->get('cid', [], 'array');
+		$template = array_shift($cid);
+
+		if (!$template)
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE_NOT_SPECIFIED'));
+
+			return;
+		}
+
+		$tBaseDir = KunenaPath::clean(KPATH_SITE . '/template');
+
+		if (!is_dir($tBaseDir . '/' . $template))
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_TEMPLATE_NOT_FOUND'));
+
+			return;
+		}
+
+		$template = KunenaPath::clean($template);
+		$this->app->setUserState('kunena.edit.templatename', $template);
+
+		$this->setRedirect(KunenaRoute::_($this->baseurl . "&layout=edit&name={$template}", false));
+	}
+
+	/**
+	 * Choose less
+	 *
+	 * @return  void
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 * @since   Kunena 2.0
+	 */
+	public function chooseLess(): void
+	{
+		$template     = $this->app->input->getArray(['cid' => '']);
+		$templatename = array_shift($template['cid']);
+		$this->app->setUserState('kunena.templatename', $templatename);
+
+		$tBaseDir = KunenaPath::clean(KPATH_SITE . '/template');
+
+		if (!is_dir($tBaseDir . '/' . $templatename . '/assets/less'))
+		{
+			$this->app->enqueueMessage(Text::_('COM_KUNENA_A_TEMPLATE_MANAGER_NO_LESS'), 'warning');
+
+			return;
+		}
+
+		$this->setRedirect(KunenaRoute::_($this->baseurl . "&layout=chooseLess", false));
+	}
+
+	/**
+	 * Choose Css
+	 *
+	 * @return  void
+	 *
+	 * @throws  Exception
+	 * @throws  null
+	 * @since   Kunena 2.0
+	 */
+	public function chooseCss(): void
+	{
+		$template     = $this->app->input->getArray(['cid' => '']);
+		$templatename = array_shift($template['cid']);
+
+		$this->app->setUserState('kunena.templatename', $templatename);
+
+		$this->setRedirect(KunenaRoute::_($this->baseurl . "&layout=chooseCss", false));
+	}
 }
